+++ source_gen/source_gen/lib/src/library.dart
@@ -2,6 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:analyzer/dart/ast/ast.dart';
+import 'package:analyzer/dart/ast/standard_resolution_map.dart';
 import 'package:analyzer/dart/element/element.dart';
 // ignore: implementation_imports
 import 'package:analyzer/src/dart/resolver/scope.dart';
@@ -23,4 +25,30 @@ class LibraryReader {
   /// that are accessible via one or more `export` directives.
   ClassElement findType(String name) =>
       _element.getType(name) ?? _namespace.get(name) as ClassElement;
+
+  /// Returns all of the declarations in this library, including the
+  /// [LibraryElement] as the first item.
+  Iterable<Element> get allElements sync* {
+    yield _element;
+    for (var cu in _element.units) {
+      for (var compUnitMember in cu.unit.declarations) {
+        yield* _getElements(compUnitMember);
+      }
+    }
+  }
+
+  Iterable<Element> _getElements(CompilationUnitMember member) {
+    if (member is TopLevelVariableDeclaration) {
+      return member.variables.variables
+          .map(resolutionMap.elementDeclaredByVariableDeclaration);
+    }
+    var element = resolutionMap.elementDeclaredByDeclaration(member);
+
+    if (element == null) {
+      print([member, member.runtimeType, member.element]);
+      throw new Exception('Could not find any elements for the provided unit.');
+    }
+
+    return [element];
+  }
 }